Move db addData function into separate module Make setup-db-generic imports addData function
